Contracts project refactoring
